Remove task manager, add simplified actor
