+++ Cargo.toml
@@ -19,7 +19,7 @@ lark-cli = { path = "components/lark-cli" }
 
 [dev-dependencies]
 env_logger = "0.6"
-languageserver-types = "0.51.1"
+languageserver-types = "0.53.0"
 lark-debug-derive = { path = "components/lark-debug-derive" }
 lark-debug-with = { path = "components/lark-debug-with" }
 lark-entity = { path = "components/lark-entity" }
