+++ components/lark-actor/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "lark-actor"
+version = "0.1.0"
+authors = ["Jonathan Turner <jonathan.d.turner@gmail.com>"]
+edition = "2018"
+
+[dependencies]
+url = "1.7"
+languageserver-types = "0.53"
