+++ components/lark-cli/Cargo.toml
@@ -21,7 +21,7 @@ typed-arena = "1.4"
 unicode-xid = "0.1.0"
 itertools = "0.8"
 text-diff = "0.4.0"
-languageserver-types = "0.51.1"
+languageserver-types = "0.53"
 flexi_logger = "0.10.3"
 pretty_env_logger = "0.3"
 salsa = "0.8"
@@ -30,6 +30,7 @@ serde = "1.0"
 serde_derive = "1.0"
 url = "1.7"
 
+lark-actor = { path = "../lark-actor" }
 lark-build = { path = "../lark-build" }
 lark-collections = { path = "../lark-collections" }
 lark-debug-derive = { path = "../lark-debug-derive" }
@@ -45,7 +46,6 @@ lark-parser = { path = "../lark-parser" }
 lark-query-system = { path = "../lark-query-system" }
 lark-span = { path = "../lark-span" }
 lark-string = { path = "../lark-string" } 
-lark-task-manager = { path = "../lark-task-manager" }
 lark-ty = { path = "../lark-ty" }
 lark-type-check = { path = "../lark-type-check" }
 lark-unify = { path = "../lark-unify" }
