+++ components/lark-language-server/Cargo.toml
@@ -8,6 +8,6 @@ edition = "2018"
 serde_json = "1.0"
 serde = "1.0"
 serde_derive = "1.0"
-languageserver-types = "0.51.1"
+languageserver-types = "0.53"
 
-lark-task-manager = { path = "../lark-task-manager" }
+lark-actor = { path = "../lark-actor" }
