+++ components/lark-parser/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2018"
 [dependencies]
 derive-new = "0.5.6"
 itertools = "0.8"
-languageserver-types = "0.51.1"
+languageserver-types = "0.53"
 lark-collections = { path = "../lark-collections" }
 lark-debug-derive = { path = "../lark-debug-derive" }
 lark-debug-with = { path = "../lark-debug-with" }
