+++ components/lark-query-system/Cargo.toml
@@ -5,8 +5,9 @@ authors = ["Jonathan Turner <jonathan.d.turner@gmail.com>"]
 edition = "2018"
 
 [dependencies]
-languageserver-types = "0.51.1"
+languageserver-types = "0.53"
 language-reporting = { git = "https://github.com/nikomatsakis/language-reporting", branch = "remove-codespan-dep" }
+lark-actor = { path = "../lark-actor" }
 lark-collections = { path = "../lark-collections" }
 lark-debug-with = { path = "../lark-debug-with" }
 lark-entity = { path = "../lark-entity" }
@@ -17,7 +18,6 @@ lark-parser = { path = "../lark-parser" }
 lark-pretty-print = { path = "../lark-pretty-print" }
 lark-span = { path = "../lark-span" }
 lark-string = { path = "../lark-string" }
-lark-task-manager = { path = "../lark-task-manager" }
 lark-ty = { path = "../lark-ty" }
 lark-type-check = { path = "../lark-type-check" }
 log = "0.4.6"
