--- components/lark-task-manager/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "lark-task-manager"
-version = "0.1.0"
-authors = ["Jonathan Turner <jonathan.d.turner@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-languageserver-types = "0.51.1"
-lark-collections = { path = "../lark-collections" }
-url = "1.7"
